separate out ybox-pkg commands that don't require state.db vs those that do
- separation of the commands that don't require state.db like list, info etc avoids
  locking the database while they are running; the calling point in pkg.py sets a default
  flag for the same (needs_state=True) which is unset for these commands so that the
  calling point can either lock the db and pass them the state object, or can skip it
- cleanly deal with failures in subprocess run in run-user-bash-cmd
- wait after sending SIGTERM to subprocesses in entrypoint.sh only if it was
  successfully sent to at least one process
- don't try to overwrite pip.conf in entrypoint-user.sh if it already exists
- deal with possible absence of pyenv in pyenv/activate (e.g. in some containers)

+++ src/ybox/conf/resources/entrypoint-user.sh
@@ -13,9 +13,11 @@ mkdir -p "$user_home/.gnupg" && chmod 0700 "$user_home/.gnupg"
 echo "keyserver $DEFAULT_GPG_KEY_SERVER" > "$user_home/.gnupg/dirmngr.conf"
 rm -f "$user_home"/.gnupg/*/*.lock
 
-echo_color "$fg_cyan" "Enabling python pip installation for $current_user" >> $status_file
-mkdir -p "$user_home/.config/pip"
-cat > "$user_home/.config/pip/pip.conf" << EOF
+if [ ! -e "$user_home/.config/pip/pip.conf" ]; then
+  echo_color "$fg_cyan" "Enabling python pip installation for $current_user" >> $status_file
+  mkdir -p "$user_home/.config/pip"
+  cat > "$user_home/.config/pip/pip.conf" << EOF
 [global]
 break-system-packages = true
 EOF
+fi
